resources folder, requirements modified, see details
added basic implementation of websocket, moved several things to a resources folder, removed pathmanager to be replaced with something else in the future (singleton or the highscoresbot api maybe). quickstart.py removed since that will be absolete once the unity highscores migration is done (the highscoresbot api should be the replacement for the databases).